requirements.txt
pandas
matplotlib
yfinance
mplfinance
python-docx